psx/vcd: test against a synthetic Video CD image, and fix UB in the EDC check
Nothing tested the real disc path. The decoder tests work on elementary
streams, the pipeline test hands synthesised sectors straight to
VCD_FeedSector, and the probe test serves control sectors through a stub
CDIF. None of them touch a CUE, a TOC, Mode 2 sector framing as CDIF
hands it over, or whether VCD_ProbeDisc finds the control sectors at
their mandated LBAs on an actual image -- and no pressed Video CD is to
hand.

tools/vcd/make_vcd.py builds one: a two-track CUE/BIN with the ISO
volume descriptors, the CD-i Bridge system identifier, INFO.VCD at
00:04:00 and ENTRIES.VCD at 00:04:01 where the specification puts them,
and the program stream in Form 2 sectors on track 2.
tools/vcd/vcd_disc.c opens it with CDIF_Open, reads the TOC, probes, and
walks track 2 with CDIF_ReadRawSector -- the same call cdc.c's sector
tap uses.

All three test streams now pass end to end: NTSC 88 pictures and 126720
audio frames, PAL 98 and 171648 at 352x288 and 25 fps, and a B-picture
stream, with the probe identifying the disc and deriving a chapter LBA
from BCD MSF that agrees with the TOC independently.

Writing the generator found four defects in it, each caught by the real
reader rejecting what a hand-rolled one accepts. CUE INDEX times are
file-relative, not absolute, so using the absolute form shifts every
track by two seconds. A Mode 2 Form 1 sector is 12+4+8+2048+4+276 with
no 8-byte reserved gap -- Mode 1 has one, Mode 2 Form 1 spends those
bytes on the subheader -- and including both makes 2360-byte sectors
that nothing reports: sector 0 reads correctly and everything after it
drifts. The EDC must be valid, because CDIF_ReadSector returns nothing
for a sector that fails edc_lec_check_and_correct, so the probe reads
zero bytes while raw reads of the same sector work, which looks like a
probe bug. And the EDC uses the CD polynomial, not the ordinary CRC-32
one.

That last point turned up a genuine discrepancy in the tree, recorded
here because it is not this commit's to fix: lec.c's calc_edc and
recover-raw.c's EDCCrc32 compute different values for the same bytes.
EDCCrc32 uses the CD polynomial and is what validates real discs;
lec.c's does not match it. lec.c's encoders are reachable only through
CDUtility.c wrappers that nothing in this core calls, which is
presumably why it has gone unnoticed.

Also fixes real undefined behaviour the harness surfaced under UBSan.
CheckEDC assembled the stored CRC with cd_frame[n] << 24 on an unsigned
char, which promotes to int and overflows for any sector whose EDC high
byte has the top bit set -- roughly half of them. Cast to uint32_t
first.

Core builds clean; all three disc tests pass under ASan and UBSan.

+++ mednafen/cdrom/recover-raw.c
@@ -69,13 +69,17 @@ void Kill_LEC_Correct(void)
 
 int CheckEDC(const unsigned char *cd_frame, bool xa_mode)
 { 
-   unsigned int expected_crc, real_crc;
+   uint32_t expected_crc, real_crc;
    unsigned int crc_base = xa_mode ? 2072 : 2064;
 
-   expected_crc = cd_frame[crc_base + 0] << 0;
-   expected_crc |= cd_frame[crc_base + 1] << 8;
-   expected_crc |= cd_frame[crc_base + 2] << 16;
-   expected_crc |= cd_frame[crc_base + 3] << 24;
+   /* Cast before shifting: cd_frame is unsigned char, which promotes to int,
+    * and shifting a value >= 0x80 left by 24 overflows a signed 32-bit int.
+    * That is undefined behaviour, and it happens for roughly half of all
+    * sectors -- any whose EDC high byte has the top bit set. */
+   expected_crc  = (uint32_t)cd_frame[crc_base + 0] << 0;
+   expected_crc |= (uint32_t)cd_frame[crc_base + 1] << 8;
+   expected_crc |= (uint32_t)cd_frame[crc_base + 2] << 16;
+   expected_crc |= (uint32_t)cd_frame[crc_base + 3] << 24;
 
    if(xa_mode) 
       real_crc = EDCCrc32(cd_frame+16, 2056);
